ContractExtension & InvoiceMonthly fake data

+++ backend/database/migrations/2025_07_24_144232_add_room_fee_to_invoices_table.php
@@ -9,7 +9,7 @@ class AddRoomFeeToInvoicesTable extends Migration
     public function up()
     {
         Schema::table('invoices', function (Blueprint $table) {
-            $table->decimal('room_fee', 10, 2)->after('year')->nullable();
+            $table->integer('room_fee')->after('year')->nullable();
         });
     }
